Remove unnecessary bus
